+++ front/src/components/dashboard/Dashboard.jsx
@@ -1,353 +1,26 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import React from 'react';
-import { Fragment, useState } from 'react';
-import { Dialog, Menu, Transition } from '@headlessui/react';
-import { ChevronDownIcon, SearchIcon } from '@heroicons/react/solid';
-import {
-  ArchiveIcon,
-  BanIcon,
-  BellIcon,
-  FlagIcon,
-  InboxIcon,
-  MenuIcon,
-  PencilAltIcon,
-  UserCircleIcon,
-  XIcon
-} from '@heroicons/react/outline';
+import Column from './Column';
+import Header from './Header';
+import LeftMenu from './LeftMenu';
 
-const user = {
-  name: 'Whitney Francis',
-  email: 'whitneyfrancis@example.com',
-  imageUrl:
-    'https://images.unsplash.com/photo-1517365830460-955ce3ccd263?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80'
-};
-const navigation = [
-  {
-    name: 'Inboxes',
-    href: '#',
-    children: [
-      { name: 'Technical Support', href: '#' },
-      { name: 'Sales', href: '#' },
-      { name: 'General', href: '#' }
-    ]
-  },
-  { name: 'Reporting', href: '#', children: [] },
-  { name: 'Settings', href: '#', children: [] }
-];
-const sidebarNavigation = [
-  { name: 'Open', href: '#', icon: InboxIcon, current: true },
-  { name: 'Archive', href: '#', icon: ArchiveIcon, current: false },
-  { name: 'Customers', href: '#', icon: UserCircleIcon, current: false },
-  { name: 'Flagged', href: '#', icon: FlagIcon, current: false },
-  { name: 'Spam', href: '#', icon: BanIcon, current: false },
-  { name: 'Drafts', href: '#', icon: PencilAltIcon, current: false }
-];
-const userNavigation = [
-  { name: 'Your Profile', href: '#' },
-  { name: 'Sign out', href: '#' }
-];
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
-}
+import { COLUMNS_INFOS } from './dashboard.constants';
 
 const Dashboard = () => {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-
   return (
-      <div className='h-full flex flex-col'>
-        {/* Top nav*/}
-        <header className='flex-shrink-0 relative h-16 bg-white flex items-center'>
-          {/* Logo area */}
-          <div className='absolute inset-y-0 left-0 md:static md:flex-shrink-0'>
-            <a
-              href='#'
-              className='flex items-center justify-center h-16 w-16 bg-indigo-500 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-600 md:w-20'
-            >
-              <img
-                className='h-8 w-auto'
-                src='https://tailwindui.com/img/logos/workflow-mark.svg?color=white'
-                alt='Workflow'
-              />
-            </a>
-          </div>
+    <div className='h-full flex flex-col'>
+      {/* Top nav*/}
+      <Header />
+      {/* Bottom section */}
+      <div className='min-h-0 flex-1 flex overflow-hidden'>
+        {/* Narrow sidebar*/}
+        <LeftMenu />
 
-          {/* Picker area */}
-          <div className='mx-auto md:hidden'>
-            <div className='relative'>
-              <label htmlFor='inbox-select' className='sr-only'>
-                Choose inbox
-              </label>
-              <select
-                id='inbox-select'
-                className='rounded-md border-0 bg-none pl-3 pr-8 text-base font-medium text-gray-900 focus:ring-2 focus:ring-indigo-600'
-                defaultValue={
-                  sidebarNavigation.find((item) => item.current).name
-                }
-              >
-                {sidebarNavigation.map((item) => (
-                  <option key={item.name}>{item.name}</option>
-                ))}
-              </select>
-              <div className='pointer-events-none absolute inset-y-0 right-0 flex items-center justify-center pr-2'>
-                <ChevronDownIcon
-                  className='h-5 w-5 text-gray-500'
-                  aria-hidden='true'
-                />
-              </div>
-            </div>
-          </div>
-
-          {/* Menu button area */}
-          <div className='absolute inset-y-0 right-0 pr-4 flex items-center sm:pr-6 md:hidden'>
-            {/* Mobile menu button */}
-            <button
-              type='button'
-              className='-mr-2 inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-600'
-              onClick={() => setMobileMenuOpen(true)}
-            >
-              <span className='sr-only'>Open main menu</span>
-              <MenuIcon className='block h-6 w-6' aria-hidden='true' />
-            </button>
-          </div>
-
-          {/* Desktop nav area */}
-          <div className='hidden md:min-w-0 md:flex-1 md:flex md:items-center md:justify-between'>
-            <div className='min-w-0 flex-1'>
-              <div className='max-w-2xl relative text-gray-400 focus-within:text-gray-500'>
-                <label htmlFor='desktop-search' className='sr-only'>
-                  Search
-                </label>
-                <input
-                  id='desktop-search'
-                  type='search'
-                  placeholder='Search'
-                  className='block w-full border-transparent pl-12 placeholder-gray-500 focus:border-transparent sm:text-sm focus:ring-0'
-                />
-                <div className='pointer-events-none absolute inset-y-0 left-0 flex items-center justify-center pl-4'>
-                  <SearchIcon className='h-5 w-5' aria-hidden='true' />
-                </div>
-              </div>
-            </div>
-            <div className='ml-10 pr-4 flex-shrink-0 flex items-center space-x-10'>
-              <nav aria-label='Global' className='flex space-x-10'>
-                <a href='#' className='text-sm font-medium text-gray-900'>
-                  Inboxes
-                </a>
-                <a href='#' className='text-sm font-medium text-gray-900'>
-                  Reporting
-                </a>
-                <a href='#' className='text-sm font-medium text-gray-900'>
-                  Settings
-                </a>
-              </nav>
-              <div className='flex items-center space-x-8'>
-                <span className='inline-flex'>
-                  <a
-                    href='#'
-                    className='-mx-1 bg-white p-1 rounded-full text-gray-400 hover:text-gray-500'
-                  >
-                    <span className='sr-only'>View notifications</span>
-                    <BellIcon className='h-6 w-6' aria-hidden='true' />
-                  </a>
-                </span>
-
-                <div className='relative inline-block text-left'>
-                  <div
-                    onClick={() => setMobileMenuOpen(true)}
-                    className='cursor-pointer bg-white rounded-full flex text-sm focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-600'
-                  >
-                    <span className='sr-only'>Open user menu</span>
-                    <img
-                      className='h-8 w-8 rounded-full'
-                      src={user.imageUrl}
-                      alt=''
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-
-          {/* Mobile menu, show/hide this `div` based on menu open/closed state */}
-          {/* <Transition.Root show={mobileMenuOpen} as={Fragment}>
-            <Dialog
-              as='div'
-              className='fixed inset-0 z-40 md:hidden'
-              onClose={setMobileMenuOpen}
-            >
-              <Transition.Child
-                as={Fragment}
-                enter='transition-opacity ease-linear duration-300'
-                enterFrom='opacity-0'
-                enterTo='opacity-100'
-                leave='transition-opacity ease-linear duration-300'
-                leaveFrom='opacity-100'
-                leaveTo='opacity-0'
-              >
-                <Dialog.Overlay className='hidden sm:block sm:fixed sm:inset-0 sm:bg-gray-600 sm:bg-opacity-75' />
-              </Transition.Child>
-
-              <Transition.Child
-                as={Fragment}
-                enter='transition ease-out duration-150 sm:ease-in-out sm:duration-300'
-                enterFrom='transform opacity-0 scale-110 sm:translate-x-full sm:scale-100 sm:opacity-100'
-                enterTo='transform opacity-100 scale-100  sm:translate-x-0 sm:scale-100 sm:opacity-100'
-                leave='transition ease-in duration-150 sm:ease-in-out sm:duration-300'
-                leaveFrom='transform opacity-100 scale-100 sm:translate-x-0 sm:scale-100 sm:opacity-100'
-                leaveTo='transform opacity-0 scale-110  sm:translate-x-full sm:scale-100 sm:opacity-100'
-              > */}
-          {mobileMenuOpen && (
-            <nav
-              className='fixed z-40 inset-0 h-full w-full bg-white sm:inset-y-0 sm:left-auto sm:right-0 sm:max-w-sm sm:w-full sm:shadow-lg'
-              aria-label='Global'
-            >
-              <div className='h-16 flex items-center justify-between px-4 sm:px-6'>
-                <a href='#'>
-                  <img
-                    className='block h-8 w-auto'
-                    src='https://tailwindui.com/img/logos/workflow-mark.svg?color=indigo&shade=500'
-                    alt='Workflow'
-                  />
-                </a>
-                <button
-                  type='button'
-                  className='-mr-2 inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-600'
-                  onClick={() => setMobileMenuOpen(false)}
-                >
-                  <span className='sr-only'>Close main menu</span>
-                  <XIcon className='block h-6 w-6' aria-hidden='true' />
-                </button>
-              </div>
-              <div className='mt-2 max-w-8xl mx-auto px-4 sm:px-6'>
-                <div className='relative text-gray-400 focus-within:text-gray-500'>
-                  <label htmlFor='mobile-search' className='sr-only'>
-                    Search all inboxes
-                  </label>
-                  <input
-                    id='mobile-search'
-                    type='search'
-                    placeholder='Search all inboxes'
-                    className='block w-full border-gray-300 rounded-md pl-10 placeholder-gray-500 focus:border-indigo-600 focus:ring-indigo-600'
-                  />
-                  <div className='absolute inset-y-0 left-0 flex items-center justify-center pl-3'>
-                    <SearchIcon className='h-5 w-5' aria-hidden='true' />
-                  </div>
-                </div>
-              </div>
-              <div className='max-w-8xl mx-auto py-3 px-2 sm:px-4'>
-                {navigation.map((item) => (
-                  <Fragment key={item.name}>
-                    <a
-                      href={item.href}
-                      className='block rounded-md py-2 px-3 text-base font-medium text-gray-900 hover:bg-gray-100'
-                    >
-                      {item.name}
-                    </a>
-                    {item.children.map((child) => (
-                      <a
-                        key={child.name}
-                        href={child.href}
-                        className='block rounded-md py-2 pl-5 pr-3 text-base font-medium text-gray-500 hover:bg-gray-100'
-                      >
-                        {child.name}
-                      </a>
-                    ))}
-                  </Fragment>
-                ))}
-              </div>
-              <div className='border-t border-gray-200 pt-4 pb-3'>
-                <div className='max-w-8xl mx-auto px-4 flex items-center sm:px-6'>
-                  <div className='flex-shrink-0'>
-                    <img
-                      className='h-10 w-10 rounded-full'
-                      src={user.imageUrl}
-                      alt=''
-                    />
-                  </div>
-                  <div className='ml-3 min-w-0 flex-1'>
-                    <div className='text-base font-medium text-gray-800 truncate'>
-                      {user.name}
-                    </div>
-                    <div className='text-sm font-medium text-gray-500 truncate'>
-                      {user.email}
-                    </div>
-                  </div>
-                  <a
-                    href='#'
-                    className='ml-auto flex-shrink-0 bg-white p-2 text-gray-400 hover:text-gray-500'
-                  >
-                    <span className='sr-only'>View notifications</span>
-                    <BellIcon className='h-6 w-6' aria-hidden='true' />
-                  </a>
-                </div>
-                <div className='mt-3 max-w-8xl mx-auto px-2 space-y-1 sm:px-4'>
-                  {userNavigation.map((item) => (
-                    <a
-                      key={item.name}
-                      href={item.href}
-                      className='block rounded-md py-2 px-3 text-base font-medium text-gray-900 hover:bg-gray-50'
-                    >
-                      {item.name}
-                    </a>
-                  ))}
-                </div>
-              </div>
-            </nav>
-          )}
-          {/* </Transition.Child>
-            </Dialog>
-          </Transition.Root> */}
-        </header>
-
-        {/* Bottom section */}
-        <div className='min-h-0 flex-1 flex overflow-hidden'>
-          {/* Narrow sidebar*/}
-          <nav
-            aria-label='Sidebar'
-            className='hidden md:block md:flex-shrink-0 md:bg-gray-800 md:overflow-y-auto'
-          >
-            <div className='relative w-20 flex flex-col p-3 space-y-3'>
-              {sidebarNavigation.map((item) => (
-                <a
-                  key={item.name}
-                  href={item.href}
-                  className={classNames(
-                    item.current
-                      ? 'bg-gray-900 text-white'
-                      : 'text-gray-400 hover:bg-gray-700',
-                    'flex-shrink-0 inline-flex items-center justify-center h-14 w-14 rounded-lg'
-                  )}
-                >
-                  <span className='sr-only'>{item.name}</span>
-                  <item.icon className='h-6 w-6' aria-hidden='true' />
-                </a>
-              ))}
-            </div>
-          </nav>
-
-          {/* Main area */}
-          <main className='min-w-0 flex-1 border-t border-gray-200 lg:flex'>
-            {/* Primary column */}
-            <section
-              aria-labelledby='primary-heading'
-              className='min-w-0 flex-1 h-full flex flex-col overflow-y-auto lg:order-last'
-            >
-              <h1 id='primary-heading' className='sr-only'>
-                Home
-              </h1>
-              column test 1
-            </section>
-
-            {/* Secondary column (hidden on smaller screens) */}
-            <aside className='hidden lg:block lg:flex-shrink-0 lg:order-first'>
-              <div className='h-full relative flex flex-col w-96 border-r border-gray-200 bg-gray-100 overflow-y-auto'>
-                Column test
-              </div>
-            </aside>
-          </main>
-        </div>
+        {/* Main area */}
+        {COLUMNS_INFOS.map(({ title, description }) => (
+          <Column title={title} description={description} />
+        ))}
       </div>
+    </div>
   );
 };
